feat: RRv7 用に設定
