+++ vite-storybook.config.ts
@@ -1,0 +1,11 @@
+import tailwindcss from "@tailwindcss/vite";
+import { defineConfig, loadEnv } from "vite";
+import tsconfigPaths from "vite-tsconfig-paths";
+
+export default defineConfig(({ mode }) => {
+  const env = loadEnv(mode, process.cwd(), "");
+  process.env = { ...process.env, ...env };
+  return {
+    plugins: [tailwindcss(), tsconfigPaths()],
+  };
+});
